+++ client/Assets/_Valgor/Scripts/Runtime/Core/SceneIds.cs
@@ -5,7 +5,7 @@ namespace Valgor.Core
         public const string Bootstrap = "Bootstrap";
         public const string Loading = "Loading";
         public const string MainMenu = "MainMenu";
-        public const string PlayerCity = "PlayerCity";
+        public const string City = "City";
         public const string WorldMap = "WorldMap";
     }
 }